Added error handling to install script
Since shell toolkit is used in the installation script, error handling
was added to ensure that it could be sourced.

Additionally, renamed the functions in the installation script since we
may want to use the 'install' command in the future.

README fixed a typo that referred to 'install' instead of 'install.sh'